+++ tase/telegram/bots/bot_commands/admin/extract_usernames_command.py
@@ -1,7 +1,7 @@
 import pyrogram
 from pydantic import Field
 
-from tase.common.utils import find_telegram_usernames
+from tase.common.preprocessing import find_telegram_usernames
 from tase.db.arangodb import graph as graph_models
 from tase.db.arangodb.graph.vertices.user import UserRole
 from tase.telegram.tasks import ExtractUsernamesTask
